requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
email-validator==2.1.1
fastapi==0.110.0
grpcio-tools==1.62.1
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
nacos-sdk-python==0.1.14
numpy==1.26.4
opencv-python==4.9.0.80
orjson==3.9.15
pip==23.2.1
protobuf==4.25.3
pydantic-core==2.16.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic==2.6.4
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
setuptools==68.1.2
sniffio==1.3.1
starlette==0.36.3
typing-extensions==4.10.0
ujson==5.9.0
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
